+++ test/formatter-edge-cases.test.ts
@@ -17,6 +17,7 @@ import { formatNpmrc } from '../src/formatters/npmrc.js'
 import { formatProperties } from '../src/formatters/properties.js'
 import { formatScreen } from '../src/formatters/screen.js'
 import { formatSsh } from '../src/formatters/ssh.js'
+import { formatSystemd } from '../src/formatters/systemd.js'
 import { formatToml } from '../src/formatters/toml.js'
 import { formatTmux } from '../src/formatters/tmux.js'
 import { formatYarnrc } from '../src/formatters/yarnrc.js'
@@ -34,6 +35,7 @@ const formatters = [
   ['tmux', formatTmux, 'set -g mouse on\r\n'],
   ['screen', formatScreen, 'startup_message off\r\n'],
   ['inputrc', formatInputrc, 'set bell-style visible\r\n'],
+  ['systemd', formatSystemd, '[Unit]\r\nDescription=Demo\r\n'],
   ['env', formatEnv, 'KEY=value\r\n'],
   ['INI', formatIni, '[section]\r\nkey=value\r\n'],
   ['properties', formatProperties, 'key=value\r\n'],